requirements.txt
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
sqlalchemy[asyncio]>=2.0.25,<3.0.0
alembic>=1.13.0,<2.0.0
jinja2>=3.1.3,<4.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt==4.0.1
python-multipart>=0.0.6,<1.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
aiosqlite>=0.19.0,<1.0.0
httpx>=0.26.0,<1.0.0
email-validator>=2.1.0,<3.0.0
pytest>=7.4.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0